chore(builder): use rspack.xxxPlugin instead of builtins.xxx
