+++ packages/builder/builder-webpack-provider/src/shared/plugin.ts
@@ -18,7 +18,7 @@ export const applyBasicPlugins = (plugins: Plugins) =>
     import('../plugins/copy').then(m => m.builderPluginCopy()),
     plugins.html(),
     plugins.image(),
-    import('../plugins/define').then(m => m.builderPluginDefine()),
+    plugins.define(),
     import('../plugins/tsLoader').then(m => m.builderPluginTsLoader()),
     import('../plugins/babel').then(m => m.builderPluginBabel()),
     import('../plugins/css').then(m => m.builderPluginCss()),
@@ -46,7 +46,7 @@ export const applyDefaultPlugins = (plugins: Plugins) =>
     plugins.wasm(),
     plugins.moment(),
     plugins.nodeAddons(),
-    import('../plugins/define').then(m => m.builderPluginDefine()),
+    plugins.define(),
     import('../plugins/progress').then(m => m.builderPluginProgress()),
     import('../plugins/minimize').then(m => m.builderPluginMinimize()),
     import('../plugins/manifest').then(m => m.builderPluginManifest()),
